database seed message and JSON package updates
