+++ database/database.js
@@ -15,8 +15,23 @@ var pledgeSchema = mongoose.Schema({
 
 var PledgeModel = mongoose.model('PledgeModel', pledgeSchema);
 
-let find = (id, callback) => {
-  PledgeModel.find().where('projectId').equals(id).exec(callback);
+let methods = {
+
+  find: (id, callback) => {
+    PledgeModel.find().where('projectId').equals(id).exec(callback);
+  },
+
+  save: (obj) => {
+    let listDocument = new PledgeModel({
+      projectId: obj.projectId,
+      pledgeInfo: obj.pledgeInfo,
+      pledgeSupport: obj.pledgeSupport
+    })
+    listDocument.save(function (err) {
+      if (err) return console.error(err);
+      console.log('Database successfully seeded')
+    });
+  }
 }
 
-module.exports.find = find;
+module.exports.methods = methods;
