+++ package.json
@@ -29,13 +29,15 @@
   "dependencies": {
     "babel-core": "^6.17.0",
     "body-parser": "^1.15.2",
+    "cors": "^2.8.4",
     "css-loader": "^0.28.9",
     "express": "^4.14.0",
     "jquery": "^3.1.1",
-    "style-loader": "^0.20.2",
+    "mongoose": "^5.0.9",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "request": "^2.75.0"
+    "request": "^2.75.0",
+    "style-loader": "^0.20.2"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
